added silent log level for test env

+++ auth/src/routes/users/user.test.ts
@@ -16,12 +16,14 @@ if (env.ENVIRONMENT !== "test") {
 const client = testClient(createApp().route("/", router));
 
 describe("user Routes", () => {
-  beforeAll(async () => {
-    await execSync(`bunx drizzle-kit push`);
+  beforeAll(() => {
+    execSync(`bunx drizzle-kit push`);
   });
 
   afterAll(async () => {
-    fs.rmSync("test.db", { force: true });
+    if (fs.existsSync("test.db")) {
+      fs.rmSync("test.db", { force: true });
+    }
   });
 
   it("get /users validates the body when creating a new user", async () => {
@@ -33,8 +35,18 @@ describe("user Routes", () => {
     expect(response.status).toBe(422);
     if (response.status === 422) {
       const json = await response.json();
-      expect(json.error.issues[0].path[0]).toBe("id");
-      expect(json.error.issues[0].message).toBe(ZOD_ERROR_MESSAGES.EXPECTED_NUMBER);
+      expect(json.error).toMatchObject({
+        issues: [
+          {
+            path: ["id"],
+            message: ZOD_ERROR_MESSAGES.EXPECTED_NUMBER,
+          },
+        ],
+      });
     }
   });
 });
+
+// describe("user handlers", () => {
+
+// });
